Refactor(mentor): checkToken.js -> authorization.js 파일명 변경
