+++ index.html
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="css/reset.css" />
     <link rel="stylesheet" href="css/variables.css" />
     <link rel="stylesheet" href="css/style.css" />
-    <script defer type="module" src="js/checkToken.js"></script>
+    <script defer type="module" src="js/authorization.js"></script>
     <title>Linkbrary</title>
   </head>
   <body>
